pyproject.toml
[project]
name = "rag"
description = ""
readme = "README.md"
requires-python = "~=3.12"
version = "0.0.0"
dependencies = [
    "pydantic-settings==2.9.1",
    "polars==1.31.0",
    "bm25s==0.2.13",
    "sentence-transformers==5.0.0",
    "torch==2.6.0",
    "transformers[torch]==4.48.3",
    "fastapi==0.116.1",
    "uvicorn[standard]==0.34.3",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true